# Avoid calling Math.pow if possible.