@tamuratak 2020-06-07
Thank you for your responses. I have edited as requested.